+++ packages/panda/src/theme/recipes/textarea.ts
@@ -8,6 +8,7 @@ export const textarea = defineRecipe({
     borderColor: 'border.default',
     borderRadius: 'l2',
     borderWidth: '1px',
+    colorPalette: 'accent',
     minWidth: 0,
     outline: 0,
     position: 'relative',
@@ -19,8 +20,8 @@ export const textarea = defineRecipe({
       cursor: 'not-allowed',
     },
     _focus: {
-      borderColor: 'border.accent',
-      boxShadow: 'accent',
+      borderColor: 'colorPalette.default',
+      boxShadow: '0 0 0 1px var(--colors-color-palette-default)',
     },
   },
   defaultVariants: {