+++ packages/panda/src/theme/semantic-tokens/colors/cyan.ts
@@ -23,4 +23,9 @@ export const cyan = {
   a10: { value: { base: '{colors.cyan.light.a10}', _dark: '{colors.cyan.dark.a10}' } },
   a11: { value: { base: '{colors.cyan.light.a11}', _dark: '{colors.cyan.dark.a11}' } },
   a12: { value: { base: '{colors.cyan.light.a12}', _dark: '{colors.cyan.dark.a12}' } },
+
+  default: { value: '{colors.cyan.9}' },
+  emphasized: { value: '{colors.cyan.10}' },
+  fg: { value: '{colors.white}' },
+  text: { value: '{colors.cyan.a11}' },
 }