+++ packages/panda/src/theme/semantic-tokens/colors/iris.ts
@@ -23,4 +23,9 @@ export const iris = {
   a10: { value: { base: '{colors.iris.light.a10}', _dark: '{colors.iris.dark.a10}' } },
   a11: { value: { base: '{colors.iris.light.a11}', _dark: '{colors.iris.dark.a11}' } },
   a12: { value: { base: '{colors.iris.light.a12}', _dark: '{colors.iris.dark.a12}' } },
+
+  default: { value: '{colors.iris.9}' },
+  emphasized: { value: '{colors.iris.10}' },
+  fg: { value: '{colors.white}' },
+  text: { value: '{colors.iris.a11}' },
 }