+++ packages/panda/src/theme/semantic-tokens/colors/lime.ts
@@ -23,4 +23,9 @@ export const lime = {
   a10: { value: { base: '{colors.lime.light.a10}', _dark: '{colors.lime.dark.a10}' } },
   a11: { value: { base: '{colors.lime.light.a11}', _dark: '{colors.lime.dark.a11}' } },
   a12: { value: { base: '{colors.lime.light.a12}', _dark: '{colors.lime.dark.a12}' } },
+
+  default: { value: '{colors.lime.9}' },
+  emphasized: { value: '{colors.lime.10}' },
+  fg: { value: '{colors.olive.light.12}' },
+  text: { value: '{colors.lime.a11}' },
 }