+++ packages/panda/src/theme/semantic-tokens/colors/red.ts
@@ -23,4 +23,9 @@ export const red = {
   a10: { value: { base: '{colors.red.light.a10}', _dark: '{colors.red.dark.a10}' } },
   a11: { value: { base: '{colors.red.light.a11}', _dark: '{colors.red.dark.a11}' } },
   a12: { value: { base: '{colors.red.light.a12}', _dark: '{colors.red.dark.a12}' } },
+
+  default: { value: '{colors.red.9}' },
+  emphasized: { value: '{colors.red.10}' },
+  fg: { value: '{colors.white}' },
+  text: { value: '{colors.red.a11}' },
 }