+++ packages/panda/src/theme/semantic-tokens/colors/teal.ts
@@ -23,4 +23,9 @@ export const teal = {
   a10: { value: { base: '{colors.teal.light.a10}', _dark: '{colors.teal.dark.a10}' } },
   a11: { value: { base: '{colors.teal.light.a11}', _dark: '{colors.teal.dark.a11}' } },
   a12: { value: { base: '{colors.teal.light.a12}', _dark: '{colors.teal.dark.a12}' } },
+
+  default: { value: '{colors.teal.9}' },
+  emphasized: { value: '{colors.teal.10}' },
+  fg: { value: '{colors.white}' },
+  text: { value: '{colors.teal.a11}' },
 }