+++ packages/panda/src/theme/semantic-tokens/colors/violet.ts
@@ -23,4 +23,9 @@ export const violet = {
   a10: { value: { base: '{colors.violet.light.a10}', _dark: '{colors.violet.dark.a10}' } },
   a11: { value: { base: '{colors.violet.light.a11}', _dark: '{colors.violet.dark.a11}' } },
   a12: { value: { base: '{colors.violet.light.a12}', _dark: '{colors.violet.dark.a12}' } },
+
+  default: { value: '{colors.violet.9}' },
+  emphasized: { value: '{colors.violet.10}' },
+  fg: { value: '{colors.white}' },
+  text: { value: '{colors.violet.a11}' },
 }