+++ packages/panda/src/theme/semantic-tokens/index.ts
@@ -1,6 +1,5 @@
 import { defineSemanticTokens } from '@pandacss/dev'
-import { match } from 'ts-pattern'
-import type { AccentColor, GrayColor, PresetOptions } from '../../types'
+import type { PresetOptions } from '../../types'
 import { colors } from './colors'
 import { createRadiiTokens } from './radii'
 import { shadows } from './shadows'
@@ -11,8 +10,8 @@ export const createSemanticTokens = (options: PresetOptions) => {
   return defineSemanticTokens({
     colors: {
       ...colors,
-      gray: createColorPalette(grayColor),
-      accent: createAccentColorPalette(accentColor),
+      gray: colors[grayColor],
+      accent: colors[accentColor],
       bg: {
         canvas: { value: '{colors.gray.1}' },
         default: { value: { base: '{colors.white}', _dark: '{colors.gray.2}' } },
@@ -32,46 +31,10 @@ export const createSemanticTokens = (options: PresetOptions) => {
         muted: { value: '{colors.gray.6}' },
         subtle: { value: '{colors.gray.4}' },
         disabled: { value: '{colors.gray.5}' },
-
         outline: { value: '{colors.gray.a9}' },
-        accent: { value: '{colors.accent.default}' },
       },
     },
     shadows,
     radii: createRadiiTokens(borderRadius),
   })
 }
-
-const createAccentColorPalette = (accentColor: AccentColor) => {
-  const tokens = match(accentColor)
-    .with('neutral', () => ({
-      default: { value: { base: '{colors.black}', _dark: '{colors.white}' } },
-      emphasized: { value: '{colors.gray.12}' },
-      fg: { value: { base: '{colors.white}', _dark: '{colors.black}' } },
-    }))
-    .with('amber', 'lime', 'mint', 'sky', 'yellow', () => ({
-      default: { value: '{colors.accent.9}' },
-      emphasized: { value: '{colors.accent.10}' },
-      fg: { value: '{colors.gray.light.12}' },
-    }))
-    .otherwise(() => ({
-      default: { value: '{colors.accent.9}' },
-      emphasized: { value: '{colors.accent.10}' },
-      fg: { value: '{colors.white}' },
-    }))
-
-  return {
-    ...createColorPalette(accentColor),
-    ...tokens,
-  }
-}
-
-const createColorPalette = (name: GrayColor | AccentColor) =>
-  Array.from({ length: 12 }, (_, i) => i + 1).reduce(
-    (acc, scale) => ({
-      ...acc,
-      [scale]: { value: `{colors.${name}.${scale}}` },
-      ['a' + scale]: { value: `{colors.${name}.a${scale}}` },
-    }),
-    {},
-  )