+++ packages/panda/CHANGELOG.md
@@ -3,6 +3,14 @@
 ### Added
 
 - Added `Kbd` component.
+- Added `accent.text` to the list of semantic tokens. This token is used do display text in interactive elements like
+  buttons and links.
+- Added support to render components with different colors. Pass the `colorPalette` prop to the component to change its
+  visual style. See the example below for more details.
+
+```tsx
+<Button colorPalette="red">Danger</Button>
+```
 
 ### Changed
 
@@ -11,6 +19,10 @@
 ### Fixed
 
 - Resolved an issue that caused the `Dialog` to show up in the wrong position when the page was scrolled.
+
+### Removed
+
+- Removed `border.accent` from the list of semantic tokens. Use `accent.default` or `colorPalette.default` instead.
 
 ## [0.24.0] - 2023-11-21